Make sample validation report a modal to be consistent
